нужно доделывать фильтр и календарь. Но тк верстку надо сдать до 9 переключаюсь на верстку
