+++ package.json
@@ -132,6 +132,7 @@
     "react-motion-drawer": "^3.1.0",
     "react-redux": "^7.2.2",
     "react-router-dom": "^5.2.0",
+    "react-select": "^3.2.0",
     "react-slick": "^0.27.13",
     "react-typed": "^1.2.0",
     "react-yandex-maps": "^4.6.0",
