+++ src/components/common/form-conclusion.jsx
@@ -1,0 +1,21 @@
+import React from 'react';
+
+const FormConclusion = () => {
+  return (
+    <div className="form-step__conclusion">
+      <div className="form-step__conclusion-info-container">
+        <div className="form-step__conclusion-info-container-text">
+          <p className="form-step__conclusion-info-text">
+            Первичный прием к Акушер-гинеколог
+          </p>
+          <p className="form-step__conclusion-info-text">
+            Мельникова Наталья Игоревна
+          </p>
+        </div>
+      </div>
+      <span className="form-step__conclusion-price">1 499 ₽</span>
+    </div>
+  );
+};
+
+export default FormConclusion;
