+++ src/components/common/form-info.jsx
@@ -1,0 +1,17 @@
+import React from 'react';
+
+const FormInfo = () => {
+  return (
+    <div className="specialist-form__form-info">
+      <p className="specialist-form__form-info-text">
+        Вы будете записаны к Акушер-гинеколог
+      </p>
+      <p className="specialist-form__form-info-text">
+        Мельникова Наталья Игоревна
+      </p>
+      <p className="specialist-form__form-info-text">09.10.2020 в 8:00</p>
+    </div>
+  );
+};
+
+export default FormInfo;
