+++ src/pages/specialists/components/filter-bottom-city.jsx
@@ -1,15 +1,21 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import Select from '../../../components/common/select';
+import { filterDublicatesObjects } from '../../utils/filter';
 
-const FilterBottomCity = ({ selectData }) => {
+const FilterBottomCity = ({ selectData, action }) => {
+  const onSelectChangeHandler = ({ value }) => {
+    action({ city: value });
+  };
+
   return (
     <div className="filter__bottom-group filter__bottom-group--city">
       <h3 className="filter__bottom-group-title">Медицинский центр</h3>
       <Select
         selectClass="filter__bottom-city-select"
-        data={selectData}
-        defaultOption="Выберите город"
+        data={filterDublicatesObjects(selectData)}
+        placeholder="Выберите город"
+        action={onSelectChangeHandler}
       />
     </div>
   );
@@ -17,6 +23,7 @@ const FilterBottomCity = ({ selectData }) => {
 
 FilterBottomCity.propTypes = {
   selectData: PropTypes.arrayOf(PropTypes.object).isRequired,
+  action: PropTypes.func.isRequired,
 };
 
 export default FilterBottomCity;
