+++ src/pages/specialists/components/filter-top.jsx
@@ -1,50 +1,87 @@
-import React from 'react';
+import React, { useState } from 'react';
 import PropTypes from 'prop-types';
 import { ReactComponent as SearchIcon } from '../../../icons/search-icon.svg';
 import Select from '../../../components/common/select';
+import SearchList from '../../../components/common/searchList';
 
-const FilterTop = ({ filter }) => {
+const FilterTop = ({ categories, names, action }) => {
+  const [inputValue, setInputValue] = useState('');
+  const [isSearchListShown, setSearchListShown] = useState(false);
+
+  const onSearchInputFocusHandler = () => {
+    setSearchListShown(true);
+  };
+
+  const onSearchInputBlurHandler = () => {
+    setSearchListShown(false);
+  };
+
+  const onSelectChangeHandler = ({ value }) => {
+    action({ category: value });
+  };
+
+  const onSearchListElementClickHandler = (value) => {
+    setInputValue(value);
+    action({ name: value });
+  };
+
   return (
     <div className="filter__top">
       <Select
         selectClass="filter__select"
-        data={filter.map((el) => ({
-          name: 'spec-filter',
+        data={categories.map((el) => ({
           value: el,
-          text: el,
+          label: el,
         }))}
-        defaultOption="Выбрать специалиста"
+        placeholder="Выбрать специалиста"
+        action={onSelectChangeHandler}
       />
-      <input
-        type="text"
-        className="search__input"
-        placeholder="Введите ФИО врача"
-        aria-label="Введите ФИО врача"
-        aria-describedby="search-filter-descr"
-      />
-      <p className="visually-hidden" id="search-descr">
-        Введите ФИО врача
-      </p>
-      <button
-        className="search__button"
-        type="button"
-        name="search-button"
-        aria-label="Кнопка поиска"
-      >
-        <SearchIcon
-          width="25px"
-          height="25px"
-          transform="scale(1.2)"
-          fill="none"
-          stroke="currentColor"
+      <div className="filter__search-container">
+        <input
+          type="text"
+          className="search__input"
+          placeholder="Введите ФИО врача"
+          aria-label="Введите ФИО врача"
+          value={inputValue}
+          onFocus={() => onSearchInputFocusHandler()}
+          onBlur={() => onSearchInputBlurHandler()}
+          onChange={(evt) => setInputValue(evt.currentTarget.value)}
+          aria-describedby="search-filter-descr"
         />
-      </button>
+        {isSearchListShown && (
+          <SearchList
+            classname="filter__search-list"
+            data={names}
+            value={inputValue}
+            action={onSearchListElementClickHandler}
+          />
+        )}
+        <p className="visually-hidden" id="search-descr">
+          Введите ФИО врача
+        </p>
+        <button
+          className="search__button"
+          type="button"
+          name="search-button"
+          aria-label="Кнопка поиска"
+        >
+          <SearchIcon
+            width="25px"
+            height="25px"
+            transform="scale(1.2)"
+            fill="none"
+            stroke="currentColor"
+          />
+        </button>
+      </div>
     </div>
   );
 };
 
 FilterTop.propTypes = {
-  filter: PropTypes.arrayOf(PropTypes.string).isRequired,
+  categories: PropTypes.arrayOf(PropTypes.string).isRequired,
+  action: PropTypes.func.isRequired,
+  names: PropTypes.arrayOf(PropTypes.string).isRequired,
 };
 
 export default FilterTop;
