+++ src/pages/specialists/components/filter.jsx
@@ -3,17 +3,22 @@ import PropTypes from 'prop-types';
 import FilterBottom from './filter-bottom';
 import FilterTop from './filter-top';
 
-const Filter = ({ filter }) => {
+const Filter = ({ filter, action }) => {
   return (
     <section className="specialists__filter filter">
-      <FilterTop filter={filter.specialistsCategrories} />
-      <FilterBottom centresSelect={filter.centers} />
+      <FilterTop
+        categories={filter.specialistsCategrories}
+        names={filter.specialistsNames}
+        action={action}
+      />
+      <FilterBottom centresSelect={filter.centers} action={action} />
     </section>
   );
 };
 
 Filter.propTypes = {
   filter: PropTypes.objectOf(PropTypes.array).isRequired,
+  action: PropTypes.func.isRequired,
 };
 
 export default Filter;
