+++ src/pages/specialists/components/specialists.jsx
@@ -1,4 +1,4 @@
-import React, { useState } from 'react';
+import React, { useEffect, useState } from 'react';
 import { connect } from 'react-redux';
 import PropTypes from 'prop-types';
 import MainContainer from '../../../containers/main-container';
@@ -13,24 +13,40 @@ import Components from '../../../constants/components';
 import { getFlatArr } from '../../utils/filter';
 
 const filterArrayByCategory = (arr, category) => {
+  if (!category) {
+    return arr;
+  }
+
   return arr.filter((el) => el.job.includes(category));
 };
 
 const filterArrayByName = (arr, name) => {
+  if (!name) {
+    return arr;
+  }
+
   return arr.filter((el) => el.name === name);
 };
 
 const filterArrayByCity = (arr, city) => {
+  if (!city) {
+    return arr;
+  }
+
   return arr.filter((el) => {
-    for (let i = 0; i < el.adresses.length; i + 1) {
-      if (el.adresses.city === city) return true;
+    for (let i = 0; i < el.adresses.length; i += 1) {
+      if (el.adresses[i].city === city) return true;
     }
     return false;
   });
 };
 
 const filterArrayByAges = (arr, ages) => {
-  return arr.filter((el) => el.ages === ages || el.ages === 2);
+  if (!ages || ages === 3) {
+    return arr;
+  }
+
+  return arr.filter((el) => el.ages === 3 || el.ages === ages);
 };
 
 const Specialists = ({ specialists }) => {
@@ -43,12 +59,32 @@ const Specialists = ({ specialists }) => {
     ages: '',
   });
 
+  const onChangeFiltersFieldsHanlder = (obj) => {
+    setCurrentFilter({ ...currentFilter, ...obj });
+  };
+
   const filter = {
     specialistsCategrories: [
-      ...new Set(getFlatArr(specialistsArr.map((el) => el.job))),
+      ...new Set(getFlatArr(specialists.map((el) => el.job))),
     ],
-    centers: getFlatArr(specialists.map((el) => getFlatArr(el.adresses))),
+    centers: getFlatArr(specialists.map((el) => el.adresses)),
+    specialistsNames: specialists.map((el) => el.name),
   };
+
+  useEffect(() => {
+    setSpecialistsArr(
+      filterArrayByAges(
+        filterArrayByCity(
+          filterArrayByCategory(
+            filterArrayByName([...specialists], currentFilter.name),
+            currentFilter.category,
+          ),
+          currentFilter.city,
+        ),
+        currentFilter.ages,
+      ),
+    );
+  }, [currentFilter]);
 
   return (
     <MainContainer>
@@ -56,7 +92,7 @@ const Specialists = ({ specialists }) => {
         <SectionInner>
           <BreadCrumbs className="specialists" />
           <h1 className="main__title">Специалисты</h1>
-          <Filter filter={filter} />
+          <Filter filter={filter} action={onChangeFiltersFieldsHanlder} />
           <SpecialistsCatalog specialists={specialistsArr} />
         </SectionInner>
       </section>
@@ -73,19 +109,15 @@ Specialists.propTypes = {
       ages: PropTypes.number,
       price: PropTypes.number,
       adresses: PropTypes.arrayOf(
-        PropTypes.arrayOf(
-          PropTypes.shape({
-            city: PropTypes.string,
-            center: PropTypes.arrayOf(
-              PropTypes.objectOf(
-                PropTypes.shape({
-                  name: PropTypes.string,
-                  adress: PropTypes.string,
-                }),
-              ),
-            ),
-          }),
-        ),
+        PropTypes.shape({
+          city: PropTypes.string,
+          center: PropTypes.arrayOf(
+            PropTypes.shape({
+              name: PropTypes.string,
+              adress: PropTypes.string,
+            }),
+          ),
+        }),
       ),
     }),
   ).isRequired,
