+++ src/sass/blocks/geo.scss
@@ -16,7 +16,7 @@
 .geo__select {
   position: relative;
   min-width: 100%;
-  min-height: 72px;
+  min-height: 100%;
   padding-left: 25px;
   border: none;
   appearance: none;
@@ -25,11 +25,33 @@
   font-size: 10px;
   outline: none;
 
+  .select__control {
+    border: none;
+    background-color: transparent !important;
+    outline: none;
+
+    &:focus {
+      outline: none;
+    }
+  }
+
+  .select__menu {
+    left: 0;
+  }
+
+  .select__menu-list {
+    max-height: 100% !important;
+  }
+
   &:hover,
   &:focus {
     background-color: transparent;
     color: $color-text;
   }
+}
+
+.geo__form {
+  height: 100%;
 }
 
 .geo__city {
@@ -55,10 +77,7 @@
 
   .section__inner {
     position: relative;
-  }
-
-  .select__content {
-    top: 73px;
+    height: 100%;
   }
 
   &:hover,
@@ -74,13 +93,14 @@
       color: $white;
     }
 
-    .select__content {
-      color: $color-text;
+    .select__single-value,
+    .select__placeholder,
+    .select__indicator {
+      color: $white;
     }
 
-    .select__title::after,
-    .select__title::before {
-      background-color: $white;
+    .select__option {
+      color: $color-text;
     }
   }
 }
@@ -130,17 +150,14 @@
   }
 
   .geo__select {
-    min-height: 67px;
+    min-height: 100%;
     padding-left: 21px;
-    margin-top: -5px;
-
-    .select__content {
-      top: 70px;
-    }
   }
 
   .geo__selection {
+    height: 72px;
     min-height: 72px;
+    justify-content: flex-start;
 
     &::after {
       top: 27px;
@@ -151,7 +168,7 @@
   }
 
   .geo__icon {
-    top: 26px;
+    top: 40%;
     left: 46px;
   }
