+++ src/sass/blocks/specialists-filter.scss
@@ -22,40 +22,61 @@
   flex-grow: 1;
   order: 2;
   margin-top: 5px;
-  background-image: linear-gradient(
-    -135deg,
-    #ff00eb 0,
-    #9a66f5 50%,
-    #2bd7ff 100%
-  );
-  color: $white;
 
-  .select__title {
-    display: flex;
-    justify-content: flex-start;
-    padding: 0;
-    padding-left: 49px;
+  .select__control {
+    background-image: linear-gradient(
+      -135deg,
+      #ff00eb 0,
+      #9a66f5 50%,
+      #2bd7ff 100%
+    );
 
-    &::before,
-    &::after {
-      background-color: $white;
+    &--is-focused {
+      box-shadow: none;
     }
   }
 
-  .select__content {
-    top: 55px;
-  }
+  .select__option {
+    background-color: $white;
+    color: $color-text;
 
-  .select__label {
-    padding-left: 49px;
-    &:hover {
+    &--is-focused {
+      background-color: transparent;
       background-image: linear-gradient(
         -135deg,
         #ff00eb 0,
         #9a66f5 50%,
         #2bd7ff 100%
       );
+      color: $white;
     }
+  }
+
+  .select__single-value {
+    color: $white;
+  }
+
+  .select__placeholder {
+    color: $white;
+  }
+
+  .select__indicator {
+    color: $white;
+  }
+}
+
+.filter__search-container {
+  position: relative;
+  display: flex;
+  flex-grow: 1;
+  flex-wrap: wrap;
+
+  .search__list {
+    padding-right: 74px;
+  }
+
+  .search__list-link {
+    border-color: $blue;
   }
 }
 
@@ -67,14 +88,13 @@
 
   .select {
     min-height: 60px;
-    border: 1px solid $blue;
     margin: 0;
     margin-right: 30px;
-    border-radius: 5px;
   }
 
-  .select__input:checked + label {
-    background-color: #fff;
+  .select__control {
+    border-color: $blue;
+    border-radius: 5px !important;
   }
 }
 
@@ -105,10 +125,6 @@
     flex-direction: row;
     flex-wrap: wrap;
     justify-content: center;
-  }
-
-  .select__content {
-    top: 59px;
   }
 }
 
@@ -169,6 +185,7 @@
     order: 0;
     margin-top: 0;
     margin-right: 10px;
+    margin-left: 0;
   }
 
   .filter__bottom {
