+++ src/sass/variables.scss
@@ -10,7 +10,7 @@ $bg-color: #fff;
 $border-color: RGBA(112, 112, 112, 0.2);
 $box-shadow: 1px 1px 22px 0 RGBA(74, 74, 74, 0.2);
 $green: #65bc29;
-$white: #f7f7f7;
+$white: #fff;
 $blue: #4faaed;
 $hover-color: #f7f7f7;
 $inactive-color: #dedede;
