+++ src/teamplate.html
@@ -1,9 +1,8 @@
 <!DOCTYPE html>
-<html lang="en">
+<html lang="ru">
   <head>
     <meta charset="utf-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1" />
-    <title>Лабдиагностика</title>
   </head>
   <body>
     <noscript>You need to enable JavaScript to run this app. </noscript>
